requirements.txt
google-auth
google-api-python-client
google-cloud-storage